added partial models - refactored into blueprint - middle of fixing PUT
